Script for Python Assignment
Hello Shubh, Bill and Jason - Attached is my python assignment as far as I am able to do it. There are two script files. The one called main.py is the one where I played with basic commands and ran the total number of months. The output is in the screen shot with a single line to show the month. The second script file...the one called main2.py is the rest of my PyBank script. I think I got most of it based on the example Bill showed us in class, but I could not produce a printed output file to show the result because I kept getting a syntax error for the output and format function at the end. And I could not fix it in time. Therefore, the answers are not printed, but the script ran fine without errors, so I hope that counts for something. Thank you.